Fix hashicorp connection tests disabled in the previous commit
Move tests from factory to the connection itself since vertx 4.3.4 moved the TLS loading code from connection creation to the first request (fetchKey)
